Inline Case files for ExceptionRaisedInUnexpectedLocationSpec
